Fix broken relative URLs
It turns out that a link to "foo" from page "/bar" goes to "/foo" whereas "foo" on "/bar/" goes to "/bar/foo". All that is to say, links inside the blog archive are broken because Firebase is configured to remove all trailing slashes. One solution is to force all URLs to have trailing slashes so that relative links work properly, but the extra redirect if someone manually types in a URL (which they presumably do without trailing slashes) really bugs me. Instead, I'm giving up a few dozen bytes of optimization which doesn't seem too bad.

Signed-off-by: Alex Saveau <saveau.alexandre@gmail.com>